fix table locked
