+++ src/query/catalog/src/lock.rs
@@ -15,15 +15,10 @@
 use std::sync::Arc;
 
 use databend_common_exception::Result;
-use databend_common_meta_app::schema::CreateLockRevReq;
-use databend_common_meta_app::schema::DeleteLockRevReq;
-use databend_common_meta_app::schema::ExtendLockRevReq;
-use databend_common_meta_app::schema::ListLockRevReq;
 use databend_common_meta_app::schema::LockKey;
 use databend_common_meta_app::schema::LockType;
 use databend_common_pipeline_core::LockGuard;
 
-use crate::catalog::Catalog;
 use crate::table_context::TableContext;
 
 #[async_trait::async_trait]
@@ -40,60 +35,3 @@ pub trait Lock: Sync + Send {
 
     async fn try_lock(&self, ctx: Arc<dyn TableContext>) -> Result<Option<LockGuard>>;
 }
-
-#[async_trait::async_trait]
-pub trait LockExt: Lock {
-    /// Return true if the table is locked.
-    async fn check_lock(&self, catalog: Arc<dyn Catalog>) -> Result<bool> {
-        let req = ListLockRevReq {
-            lock_key: self.gen_lock_key(),
-        };
-        let reply = catalog.list_lock_revisions(req).await?;
-        Ok(!reply.is_empty())
-    }
-
-    fn gen_create_lock_req(
-        &self,
-        user: String,
-        node: String,
-        query_id: String,
-        expire_secs: u64,
-    ) -> CreateLockRevReq {
-        CreateLockRevReq {
-            lock_key: self.gen_lock_key(),
-            user,
-            node,
-            query_id,
-            expire_secs,
-        }
-    }
-
-    fn gen_list_lock_req(&self) -> ListLockRevReq {
-        ListLockRevReq {
-            lock_key: self.gen_lock_key(),
-        }
-    }
-
-    fn gen_delete_lock_req(&self, revision: u64) -> DeleteLockRevReq {
-        DeleteLockRevReq {
-            lock_key: self.gen_lock_key(),
-            revision,
-        }
-    }
-
-    fn gen_extend_lock_req(
-        &self,
-        revision: u64,
-        expire_secs: u64,
-        acquire_lock: bool,
-    ) -> ExtendLockRevReq {
-        ExtendLockRevReq {
-            lock_key: self.gen_lock_key(),
-            revision,
-            acquire_lock,
-            expire_secs,
-        }
-    }
-}
-
-impl<T: ?Sized> LockExt for T where T: Lock {}
