+++ src/query/service/src/interpreters/interpreter_table_recluster.rs
@@ -16,7 +16,6 @@ use std::sync::Arc;
 use std::time::Duration;
 use std::time::SystemTime;
 
-use databend_common_catalog::lock::LockExt;
 use databend_common_catalog::plan::Filters;
 use databend_common_catalog::plan::PushDownInfo;
 use databend_common_catalog::table::TableExt;
@@ -42,6 +41,7 @@ use log::warn;
 
 use crate::interpreters::Interpreter;
 use crate::interpreters::InterpreterClusteringHistory;
+use crate::locks::LockExt;
 use crate::locks::LockManager;
 use crate::pipelines::executor::ExecutorSettings;
 use crate::pipelines::executor::PipelineCompleteExecutor;
@@ -128,7 +128,7 @@ impl Interpreter for ReclusterTableInterpreter {
 
             // check if the table is locked.
             let table_lock = LockManager::create_table_lock(table_info.clone())?;
-            if table_lock.check_lock(catalog.clone()).await? {
+            if !table_lock.wait_lock_expired(catalog.clone()).await? {
                 return Err(ErrorCode::TableAlreadyLocked(format!(
                     "table '{}' is locked, please retry recluster later",
                     self.plan.table
