+++ tests/sqllogictests/suites/base/09_fuse_engine/09_0016_remote_alter_recluster.test
@@ -8,9 +8,6 @@ statement ok
 USE db_09_0016
 
 statement ok
-set enable_table_lock = 0
-
-statement ok
 create table t1(a int, b int) cluster by(a+1)
 
 statement ok
@@ -128,6 +125,3 @@ DROP DATABASE db_09_0016
 
 statement ok
 set max_block_size = 65536
-
-statement ok
-set enable_table_lock = 1
