+++ tests/sqllogictests/suites/base/09_fuse_engine/09_0019_issue_8831.test
@@ -8,9 +8,6 @@ statement ok
 USE issue_8130
 
 statement ok
-set enable_table_lock = 0
-
-statement ok
 create table t(a int)
 
 statement ok
@@ -46,9 +43,6 @@ query I
 select count(1) from t1
 ----
 7
-
-statement ok
-set enable_table_lock = 1
 
 statement ok
 DROP TABLE t;
