+++ tests/sqllogictests/suites/base/09_fuse_engine/09_0020_analyze.test
@@ -8,9 +8,6 @@ statement ok
 USE db_09_0020
 
 statement ok
-set enable_table_lock = 0
-
-statement ok
 create table t(a uint64)
 
 statement ok
@@ -115,9 +112,6 @@ query T
 select * from fuse_statistic('db_09_0020', 't')
 ----
 (0,2);
-
-statement ok
-set enable_table_lock = 1
 
 statement ok
 DROP TABLE t
